refactor - isolate code for unittest
